sua lai loi BE

+++ src/main/java/org/aptech/backendmypham/services/BookingService.java
@@ -5,6 +5,6 @@ import org.aptech.backendmypham.dto.BookingDTO;
 import java.time.Instant;
 
 public interface BookingService {
-    public boolean isStaffAvailable(Integer userId, Instant requestedStartTime, Integer durationMinutes, Long appointmentIdToExclude);
-    public boolean isStaffAvailable(Integer userId, Instant requestedStartTime, Integer durationMinutes);
+    public boolean isStaffAvailable(Long userId, Instant requestedStartTime, Integer durationMinutes, Long appointmentIdToExclude);
+    public boolean isStaffAvailable(Long userId, Instant requestedStartTime, Integer durationMinutes);
 }
